refactor: move sinter tests into test_mle.py and test_mld.py
Removed standalone test_sinter.py. SinterGurobiDecoder tests now
live in test_mle.py, SinterTableDecoder tests in test_mld.py.
