+++ test/decoders/test_mld.py
@@ -1,7 +1,11 @@
+import math
+
 import numpy as np
+import pytest
+import sinter
 import stim
 
-from bloqade.decoders import TableDecoder
+from bloqade.decoders import SinterTableDecoder, TableDecoder
 from bloqade.decoders._decoders.base import BaseDecoder
 from bloqade.decoders._decoders.mld import (
     det_obs_shots_to_counts,
@@ -212,3 +216,145 @@ def test_single_shot_decode():
     result = decoder.decode(np.array([0, 1], dtype=bool))
     assert result.ndim == 1
     assert np.array_equal(result, np.array([True]))
+
+
+# --- SinterTableDecoder tests ---
+
+
+def _simple_dem():
+    return stim.DetectorErrorModel(
+        """
+        error(0.1) D0 L0
+        error(0.1) D1 L0
+        """
+    )
+
+
+def _repetition_circuit():
+    return stim.Circuit(
+        """
+        R 0 1 2
+        X_ERROR(0.1) 0 1 2
+        MZZ 0 1
+        DETECTOR rec[-1]
+        MZZ 1 2
+        DETECTOR rec[-1]
+        M 0 1 2
+        OBSERVABLE_INCLUDE(0) rec[-1] rec[-2] rec[-3]
+    """
+    )
+
+
+def _repetition_dem():
+    return _repetition_circuit().detector_error_model(
+        decompose_errors=False, approximate_disjoint_errors=True
+    )
+
+
+def _pack_dets(det_shots: np.ndarray) -> np.ndarray:
+    return np.packbits(
+        det_shots.astype(np.uint8), axis=1, bitorder="little"
+    )
+
+
+def _unpack_obs(packed_obs: np.ndarray, num_obs: int) -> np.ndarray:
+    unpacked = np.unpackbits(packed_obs, axis=1, bitorder="little")
+    return unpacked[:, :num_obs].astype(bool)
+
+
+def test_sinter_table_is_sinter_decoder():
+    decoder = SinterTableDecoder()
+    assert isinstance(decoder, sinter.Decoder)
+
+
+def test_sinter_table_compile_returns_compiled_decoder():
+    dem = _simple_dem()
+    decoder = SinterTableDecoder()
+    compiled = decoder.compile_decoder_for_dem(dem=dem)
+    assert isinstance(compiled, sinter.CompiledDecoder)
+
+
+def test_sinter_table_decode_shape_and_dtype():
+    dem = _repetition_dem()
+    decoder = SinterTableDecoder()
+    compiled = decoder.compile_decoder_for_dem(dem=dem)
+
+    det_shots = np.array([[1, 0], [0, 1], [0, 0]], dtype=bool)
+    packed_dets = _pack_dets(det_shots)
+
+    result = compiled.decode_shots_bit_packed(
+        bit_packed_detection_event_data=packed_dets
+    )
+
+    num_obs_bytes = math.ceil(dem.num_observables / 8)
+    assert result.dtype == np.uint8
+    assert result.shape == (3, num_obs_bytes)
+
+
+def test_sinter_table_decode_correctness():
+    dem = _repetition_dem()
+    decoder = SinterTableDecoder()
+    compiled = decoder.compile_decoder_for_dem(dem=dem)
+
+    det_shots = np.array(
+        [[1, 0], [1, 1], [0, 1], [0, 0]], dtype=bool
+    )
+    packed_dets = _pack_dets(det_shots)
+
+    result = compiled.decode_shots_bit_packed(
+        bit_packed_detection_event_data=packed_dets
+    )
+    obs_predictions = _unpack_obs(result, dem.num_observables)
+
+    expected = np.array([[True], [True], [True], [False]])
+    assert np.array_equal(obs_predictions, expected)
+
+
+def test_sinter_table_no_error_syndrome():
+    dem = _simple_dem()
+    decoder = SinterTableDecoder()
+    compiled = decoder.compile_decoder_for_dem(dem=dem)
+
+    det_shots = np.zeros((1, 2), dtype=bool)
+    packed_dets = _pack_dets(det_shots)
+
+    result = compiled.decode_shots_bit_packed(
+        bit_packed_detection_event_data=packed_dets
+    )
+    obs_predictions = _unpack_obs(result, dem.num_observables)
+
+    assert np.array_equal(obs_predictions, np.array([[False]]))
+
+
+def test_sinter_table_custom_num_shots():
+    decoder = SinterTableDecoder(num_shots=10000)
+    assert decoder.num_shots == 10000
+
+
+def test_sinter_table_default_num_shots():
+    decoder = SinterTableDecoder()
+    assert decoder.num_shots == 2**26
+
+
+@pytest.mark.slow
+def test_sinter_collect_table():
+    circuit = _repetition_circuit()
+    tasks = [
+        sinter.Task(
+            circuit=circuit,
+            decoder="table_mld",
+            json_metadata={"d": 3},
+        ),
+    ]
+    stats = sinter.collect(
+        num_workers=1,
+        tasks=tasks,
+        custom_decoders={
+            "table_mld": SinterTableDecoder(num_shots=100000)
+        },
+        max_shots=100,
+    )
+    assert len(stats) == 1
+    assert stats[0].shots == 100
+    assert stats[0].errors <= stats[0].shots
+    assert stats[0].errors < 50
